+++ changelog.md
@@ -11,6 +11,9 @@
   * A new "damaging" enchantment that increases attack speed
 * Gold Tools and Sword now have +2 damage (similar to iron tools)
 * Chainmail armor now gives toughness
+* Mobs now drop 25% more experience
+* Melee protection now increases attack speed by 2.5% per level
+* Removed sharpness
 * Forge Hammers can no longer be used as weapons
 * If you have a tool in your hand while in creative you now break blocks faster